refactor(dashboardeditor): move the default function to an external file
